requirement.txt
ultralytics
flask
APIFlask
python-dotenv
Pillow